Static draco worker (#7473)
* Avoid using new Blob(...) in DRACOLoader.ts and make worker static
* Add google-analytics to img-src
* Ignore draco worker during linting
* Merge branch 'master' into static-draco-worker
