+++ conf/application.conf
@@ -39,7 +39,7 @@ play {
       connect-src = "'self' ws://localhost:9002 app.olvy.co api.github.com gist.github.com *.google-analytics.com api.airbrake.io"
       frame-src = "'self' webknossos.olvy.co"
       # *.voxelytics.com is needed for images embedded in Voxelytics reports.
-      img-src = "'self' data: *.googletagmanager.com *.voxelytics.com"
+      img-src = "'self' data: *.googletagmanager.com *.google-analytics.com *.voxelytics.com"
       media-src = "*.webknossos.org"
       font-src = "'self' data:"
     }
